+++ test/test_quantization.py
@@ -76,6 +76,7 @@ try:
     from quantization.fx.test_subgraph_rewriter import TestSubgraphRewriter  # noqa: F401
 # Quantization for PyTorch 2.0 Export path
     from quantization.fx.test_quantize_pt2e import TestQuantizePT2E  # noqa: F401
+    from quantization.fx.test_quantize_pt2e import TestQuantizePT2EFX  # noqa: F401
     from quantization.fx.test_quantize_pt2e import TestQuantizePT2EModels  # noqa: F401
     from quantization.fx.test_quantize_pt2e import TestQuantizePT2EX86Inductor  # noqa: F401
 except ImportError:
